requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.27.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
tabulate>=0.9.0
matplotlib>=3.8.0
